chore: move eip7251 impl
